Fixed the functionality of the game
In my original program I had war drawing two face down cards and
playing the third, rather than one face down and playing the second.
This is fixed, as well as the instance when a player runs out of cards
during the war, which before made the program crash but now terminates
the loop and declares the other player the victor. Some set methods as
added in Card, static methods fixed in War to add at end only. Most of
comments are added.