This is the Final working version of the game before the deadline
The gui is working very well in terms of functionality. Ace is high,
the deck counts display the correct amounts right when the flip
occurs-- the decks have already been altered based on who won when the
card appears. Javadocs are generated and submitted in this commit.
JOptionPane displays the winners accordingly.